backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic[email]==2.9.2
openai==1.54.3
weaviate-client==4.9.3
exa-py==1.9.1
resend==2.5.0
python-dotenv==1.0.1
